CHANGE: Move VirtualMouseInput into input system DLL.
