--- Assets/Samples/GamepadMouseCursor/Unity.InputSystem.VirtualMouseInput.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c0865dd949f15ba42856c0c5582f2f47
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
