--- Assets/Tests/Samples/GamepadMouseCursorTests.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 37dd119f0a8c48d99a7bc127d6d05626
-timeCreated: 1579708580
