+++ Assets/Tests/Samples/Unity.InputSystem.Samples.Tests.asmdef
@@ -7,8 +7,7 @@
         "GUID:27619889b8ba8c24980f49ee34dbb44a",
         "GUID:0acc523941302664db1f4e527237feb3",
         "GUID:043547d79c83ab74e9bcfef914059964",
-        "GUID:157556569ac93a94fb6a763f50476339",
-        "GUID:c0865dd949f15ba42856c0c5582f2f47"
+        "GUID:157556569ac93a94fb6a763f50476339"
     ],
     "optionalUnityReferences": [],
     "includePlatforms": [],
