+++ Packages/com.unity.inputsystem/InputSystem/Plugins/UI/VirtualMouseInput.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c599c40b1335dec43b58eeec7d856c63
+guid: 424b6720ff5999446843ba414e58a907
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
